made adding concepts shared code
